move `GetCommonErrorBody` to convert_types.go
